+++ ui/components/HomeHero/HomeHero.js
@@ -6,7 +6,7 @@ const HomeHero = () => {
     <div className="c-home-hero">
       <div className="container c-home-hero__container pt-3 pt-sm-5 pb-0 pb-sm-5">
         <div className="card c-home-hero__card">
-          <StartForm/>
+          <StartForm />
         </div>
       </div>
     </div>