+++ ui/components/LogoIdea/LogoIdea.js
@@ -1,12 +1,12 @@
 import React from "react";
 import logoLBA from "../../public/images/logo-noir-seul.svg";
-import { push } from "connected-next-router";
-import { useDispatch, useSelector } from "react-redux";
+import { useRouter } from "next/router";
+import { ParameterContext } from "../../context/ParameterContextProvider";
 
 const LogoIdea = () => {
-  const dispatch = useDispatch();
+  const router = useRouter();
 
-  const { widgetParameters } = useSelector((state) => state.trainings);
+  const { widgetParameters } = React.useContext(ParameterContext);
 
   const goToLbaHome = (e) => {
     if (widgetParameters) {
@@ -19,7 +19,7 @@ const LogoIdea = () => {
       }
     } else {
       e.preventDefault();
-      dispatch(push({ pathname: "/" }));
+      router.push("/");
     }
   };