+++ ui/components/SearchForTrainingsAndJobs/components/MapListSwitchButton.js
@@ -1,10 +1,12 @@
 import React from "react";
-import { useSelector } from "react-redux";
 import toggleList from "public/images/icons/toggleList.svg";
 import toggleCard from "public/images/icons/toggleCard.svg";
+import { SearchResultContext } from "../../../context/SearchResultContextProvider";
+import { DisplayContext } from "../../../context/DisplayContextProvider";
 
 const MapListSwitchButton = ({ showResultMap, showResultList }) => {
-  const { visiblePane, hasSearch } = useSelector((state) => state.trainings);
+  const { hasSearch } = React.useContext(SearchResultContext);
+  const { visiblePane } = React.useContext(DisplayContext);
 
   if (visiblePane === "resultList") {
     if (hasSearch)