+++ ui/components/SearchForTrainingsAndJobs/components/MapPopup.js
@@ -1,6 +1,5 @@
 import React, { useEffect } from "react";
-import { setSelectedMapPopupItem, setSelectedItem } from "../../../store/actions";
-import { useDispatch } from "react-redux";
+
 import { getJobAddress } from "../../../utils/jobs";
 import { logError } from "../../../utils/tools";
 import { ErrorMessage } from "../..";
@@ -9,17 +8,15 @@ import { setSelectedMarker } from "../../../utils/mapTools";
 import bookIcon from "public/images/icons/book.svg";
 import jobIcon from "public/images/icons/job.svg";
 
-const MapPopup = ({ type, item, handleSelectItem }) => {
-  const dispatch = useDispatch();
-
+const MapPopup = ({ type, item, handleSelectItem, setSelectedItem, setSelectedMapPopupItem }) => {
   const openItemDetail = (item) => {
-    dispatch(setSelectedItem(item));
+    setSelectedItem(item);
     setSelectedMarker(item);
     handleSelectItem(item);
   };
 
   useEffect(() => {
-    dispatch(setSelectedMapPopupItem(item));
+    setSelectedMapPopupItem(item);
   }, []);
 
   const getContent = () => {
@@ -31,16 +28,21 @@ const MapPopup = ({ type, item, handleSelectItem }) => {
           <div className="c-mapbox-container">
             <div className="ml-3 my-3">
               <img className="cardIcon mr-2" src={jobIcon} alt="" />
-              <span className="mapboxPopupTitle">Opportunité<span className={`${list.length > 1 ? '' : 'd-none'}`}>s</span> d'emploi : </span>
+              <span className="mapboxPopupTitle">
+                Opportunité<span className={`${list.length > 1 ? "" : "d-none"}`}>s</span> d'emploi :{" "}
+              </span>
             </div>
-            <div className="c-mapbox-address mx-3 my-2 mb-3">
-              {getJobAddress(list[0])}
-            </div>
+            <div className="c-mapbox-address mx-3 my-2 mb-3">{getJobAddress(list[0])}</div>
             <div className="c-mapbox-bg">
               <div className="ml-3">
                 <ul className="c-mapbox-list">
                   {list.map((job, idx) => (
-                    <li className={`c-mapbox-list-item ${idx === list.length - 1 ? 'is-last' : ''} ${idx === 0 ? 'is-first' : ''}`} key={idx}>
+                    <li
+                      className={`c-mapbox-list-item ${idx === list.length - 1 ? "is-last" : ""} ${
+                        idx === 0 ? "is-first" : ""
+                      }`}
+                      key={idx}
+                    >
                       <button
                         className={`c-mapboxpopup--link gtmSavoirPlus gtm${capitalizeFirstLetter(job.ideaType)} gtmMap`}
                         onClick={() => openItemDetail(job)}
@@ -48,9 +50,7 @@ const MapPopup = ({ type, item, handleSelectItem }) => {
                         {job.title}
                       </button>
                       {job.ideaType === "peJob" && job?.company?.name ? (
-                        <span className='c-mapbox-companyname'>
-                          - {job.company.name}
-                        </span>
+                        <span className="c-mapbox-companyname">- {job.company.name}</span>
                       ) : (
                         ""
                       )}
@@ -69,9 +69,7 @@ const MapPopup = ({ type, item, handleSelectItem }) => {
               <span className="mapboxPopupTitle">Formations : </span>
             </div>
             <div className="mapboxPopupPlace mx-3 my-2">{list[0].company.name}</div>
-            <div className="mapboxPopupAddress mx-3 my-2 mb-3">
-              {list[0].place.fullAddress}
-            </div>
+            <div className="mapboxPopupAddress mx-3 my-2 mb-3">{list[0].place.fullAddress}</div>
             <div className="mapboxPopupBg">
               <div className="">
                 <div className="ml-2">
@@ -101,7 +99,7 @@ const MapPopup = ({ type, item, handleSelectItem }) => {
       );
     }
   };
-  
+
   const getTrainings = (list) => {
     let result = (
       <>
@@ -111,7 +109,7 @@ const MapPopup = ({ type, item, handleSelectItem }) => {
               <button
                 className={`c-mapboxpopup--link gtmSavoirPlus gtmFormation gtmMap`}
                 onClick={() => openItemDetail(training)}
-                >
+              >
                 {training.title ? training.title : training.longTitle}
               </button>
             </span>