+++ ui/components/WidgetHeader/WidgetHeader.js
@@ -1,15 +1,16 @@
-import React from "react";
+import React, { useContext } from "react";
 import HeaderForm from "../../components/HeaderForm/HeaderForm";
 import LogoIdea from "../../components/LogoIdea/LogoIdea";
 import { useRouter } from "next/router";
 import { includes } from "lodash";
-import { useSelector } from "react-redux";
 
 import { Row } from "reactstrap";
+import { SearchResultContext } from "../../context/SearchResultContextProvider";
 
 const WidgetHeader = ({ handleSearchSubmit, isHome }) => {
   const router = useRouter();
-  const { selectedItem } = useSelector((state) => state.trainings);
+
+  const { selectedItem } = useContext(SearchResultContext);
 
   let additionalClassName = selectedItem && includes(router.asPath, "page=fiche") ? "detail" : "global";