+++ ui/components/footer.js
@@ -1,15 +1,10 @@
 import React from "react";
 import { Row, Col, Container } from "reactstrap";
 import ExternalLink from "./externalLink";
-import { push } from "connected-next-router";
-import { useDispatch } from "react-redux";
+import { useRouter } from "next/router";
 
-/*
- Different kind of navigation are available here :
- https://raw.githubusercontent.com/danielr18/connected-next-router/master/examples/basic/components/navigation.js
-*/
 const Footer = (props) => {
-  const dispatch = useDispatch();
+  const router = useRouter();
 
   return (
     <>
@@ -28,10 +23,10 @@ const Footer = (props) => {
               <div>La bonne alternance. Trouvez votre alternance.</div>
               <div className="mt-4">La bonne alternance est proposée par les services suivants :</div>
               <div className="mt-4 c-footer-official-links">
-                <ExternalLink
-                  className="c-footer-official-link"
-                  url="https://legifrance.gouv.fr"
-                  title="legifrance.gouv.fr"
+                <ExternalLink 
+                  className="c-footer-official-link" 
+                  url="https://pole-emploi.fr" 
+                  title="pole-emploi.fr" 
                 />
                 <ExternalLink
                   className="c-footer-official-link"
@@ -43,8 +38,11 @@ const Footer = (props) => {
                   url="https://service-public.fr"
                   title="service-public.fr"
                 />
-                <ExternalLink className="c-footer-official-link" url="https://data.gouv.fr" title="data.gouv.fr" />
-                <ExternalLink className="c-footer-official-link" url="https://pole-emploi.fr" title="pole-emploi.fr" />
+                <ExternalLink 
+                  className="c-footer-official-link" 
+                  url="https://data.gouv.fr" 
+                  title="data.gouv.fr" 
+                />
               </div>
             </Col>
           </Row>
@@ -60,7 +58,7 @@ const Footer = (props) => {
                     className="c-footer-links__link c-footer-smallword pr-3"
                     onClick={(e) => {
                       e.preventDefault();
-                      dispatch(push({ pathname: "/mentions-legales" }));
+                      router.push("/mentions-legales");
                     }}
                     href="/mentions-legales"
                   >
@@ -73,7 +71,7 @@ const Footer = (props) => {
                     className="c-footer-links__link c-footer-smallword"
                     onClick={(e) => {
                       e.preventDefault();
-                      dispatch(push({ pathname: "/cgu" }));
+                      router.push("/cgu");
                     }}
                     href="/cgu"
                   >
@@ -86,7 +84,7 @@ const Footer = (props) => {
                     className="c-footer-links__link c-footer-smallword"
                     onClick={(e) => {
                       e.preventDefault();
-                      dispatch(push({ pathname: "/cookies" }));
+                      router.push("/cookies");
                     }}
                     href="/cookies"
                   >
@@ -99,7 +97,7 @@ const Footer = (props) => {
                     className="c-footer-links__link c-footer-smallword"
                     onClick={(e) => {
                       e.preventDefault();
-                      dispatch(push({ pathname: "/stats" }));
+                      router.push("/stats");
                     }}
                     href="/stats"
                   >
@@ -112,7 +110,7 @@ const Footer = (props) => {
                     className="c-footer-links__link c-footer-smallword"
                     onClick={(e) => {
                       e.preventDefault();
-                      dispatch(push({ pathname: "/faq" }));
+                      router.push("/faq");
                     }}
                     href="/faq"
                   >
@@ -125,7 +123,7 @@ const Footer = (props) => {
                     className="c-footer-links__link c-footer-smallword"
                     onClick={(e) => {
                       e.preventDefault();
-                      dispatch(push({ pathname: "/contact" }));
+                      router.push("/contact");
                     }}
                     href="/contact"
                   >
@@ -138,7 +136,7 @@ const Footer = (props) => {
                     className="c-footer-links__link c-footer-smallword"
                     onClick={(e) => {
                       e.preventDefault();
-                      dispatch(push({ pathname: "/metiers" }));
+                      router.push("/metiers");
                     }}
                     href="/metiers"
                   >
@@ -151,7 +149,7 @@ const Footer = (props) => {
                     className="c-footer-links__link c-footer-smallword"
                     onClick={(e) => {
                       e.preventDefault();
-                      dispatch(push({ pathname: "/a-propos" }));
+                      router.push("/a-propos");
                     }}
                     href="/a-propos"
                   >
@@ -164,7 +162,7 @@ const Footer = (props) => {
                     className="c-footer-links__link c-footer-smallword"
                     onClick={(e) => {
                       e.preventDefault();
-                      dispatch(push({ pathname: "/developpeurs" }));
+                      router.push("/developpeurs");
                     }}
                     href="/developpeurs"
                   >