+++ ui/components/navigation.js
@@ -1,15 +1,9 @@
 import React, { useState } from "react";
-import { Collapse, Navbar, NavbarToggler, NavbarBrand, Nav, NavItem, NavLink } from "reactstrap";
+import { Collapse, Navbar, NavbarToggler, NavbarBrand, Nav, NavItem } from "reactstrap";
+import { useRouter } from "next/router";
 
-import { push } from "connected-next-router";
-import { useDispatch } from "react-redux";
-
-/*
- Different kind of navigation are available here :
- https://raw.githubusercontent.com/danielr18/connected-next-router/master/examples/basic/components/navigation.js
-*/
 const Navigation = (props) => {
-  const dispatch = useDispatch();
+  const router = useRouter();
 
   const [isOpen, setIsOpen] = useState(false);
 
@@ -23,7 +17,13 @@ const Navigation = (props) => {
       <Navbar expand="lg" className="navbar-light">
         <div className="container">
           <NavbarBrand href="/">
-            <img src="/images/marianne.svg" alt="Logo république française" width="162" height="78" className="c-marianne-header" />
+            <img
+              src="/images/marianne.svg"
+              alt="Logo république française"
+              width="162"
+              height="78"
+              className="c-marianne-header"
+            />
             <img src="/images/logo_lba.svg" alt="Logo LBA" className="c-navbar-brand-img" width="110" height="76" />
           </NavbarBrand>
           <NavbarToggler onClick={toggle} />
@@ -34,7 +34,7 @@ const Navigation = (props) => {
                   className="nav-link"
                   onClick={(e) => {
                     e.preventDefault();
-                    dispatch(push({ pathname: "/acces-recruteur" }));
+                    router.push("/acces-recruteur");
                   }}
                   href="/acces-recruteur"
                 >
@@ -42,9 +42,16 @@ const Navigation = (props) => {
                 </a>
               </NavItem>
               <NavItem className="ml-lg-5">
-                <NavLink href="/organisme-de-formation" className="ml-1">
-                  Organisme de formation
-                </NavLink>
+                <a
+                  className="nav-link"
+                  onClick={(e) => {
+                    e.preventDefault();
+                    router.push("/organisme-de-formation");
+                  }}
+                  href="/organisme-de-formation"
+                >
+                  <span className="ml-1">Organisme de formation</span>
+                </a>
               </NavItem>
             </Nav>
           </Collapse>