+++ ui/pages/index.js
@@ -5,7 +5,8 @@ import HowTo from "components/HowTo";
 import HomeReview from "components/HomeReview";
 import { initParametersFromQuery } from "services/config";
 import Footer from "components/footer";
-import { useDispatch } from "react-redux";
+import { useRouter } from "next/router";
+
 import ScrollToTop from "components/ScrollToTop";
 import howtocircle1 from "public/images/howtocircle1.svg";
 import howtocircle2 from "public/images/howtocircle2.svg";
@@ -15,12 +16,15 @@ import howtocircle5 from "public/images/howtocircle5.svg";
 import axios from "axios";
 import csvToArray from "utils/csvToArray.js";
 import { some } from "lodash";
+import { ParameterContext } from "../context/ParameterContextProvider";
 
 const Home = (props) => {
-  const dispatch = useDispatch();
+  const router = useRouter();
+
+  const parameterContext = React.useContext(ParameterContext);
 
   useEffect(() => {
-    initParametersFromQuery({ dispatch, shouldPush: "shouldPushPathname" });
+    initParametersFromQuery({ router, shouldPush: "shouldPushPathname", parameterContext });
   }, []);
 
   return (