+++ server/CHANGELOG.md
@@ -1,5 +1,16 @@
 # CHANGELOG
 
+## 1.19.0 [22/05/2022]
+[LBA6-68] Remplacement redux par API context
+[LBA6-36] Fix effect de rechargement sur lien Organisme de formation
+
+## 1.18.12 [17/05/2022]
+[LBA6-132] Corrections en-têtes des fiches détails
+[LBA6-131] Scrolling vers les offres cassé
+[LBA6-135] Restauration info métier Matcha
+[LBA6-11] MAJ email envoyé au recruteur lors d'une candidature Matcha
+[LBA6-12] MAJ email envoyé au recruteur lors d'une candidature spontanée
+
 ## 1.18.11 [11/05/2022]
 [LBA6-32] Contrôle des emails des candidats : pas d'email temporaire
 [LBA6-110] Récupération du flag établissement mandataire